Ajustei os botões

+++ src/bean/CvsVendasjoias.java
@@ -2,6 +2,7 @@ package bean;
 // Generated 09/12/2025 21:39:33 by Hibernate Tools 4.3.1
 
 
+import java.math.BigDecimal;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -23,8 +24,8 @@ public class CvsVendasjoias  implements java.io.Serializable {
      private int cvsIdVendasJoias;
      private CvsJoias cvsJoias;
      private CvsVendas cvsVendas;
-     private Double cvsValorUnidade;
-     private int cvsQuant;
+     private BigDecimal cvsValorUnidade;
+     private Integer cvsQuant;
 
     public CvsVendasjoias() {
     }
@@ -33,7 +34,7 @@ public class CvsVendasjoias  implements java.io.Serializable {
     public CvsVendasjoias(int cvsIdVendasJoias) {
         this.cvsIdVendasJoias = cvsIdVendasJoias;
     }
-    public CvsVendasjoias(int cvsIdVendasJoias, CvsJoias cvsJoias, CvsVendas cvsVendas, Double cvsValorUnidade, int cvsQuant) {
+    public CvsVendasjoias(int cvsIdVendasJoias, CvsJoias cvsJoias, CvsVendas cvsVendas, BigDecimal cvsValorUnidade, Integer cvsQuant) {
        this.cvsIdVendasJoias = cvsIdVendasJoias;
        this.cvsJoias = cvsJoias;
        this.cvsVendas = cvsVendas;
@@ -75,21 +76,21 @@ public class CvsVendasjoias  implements java.io.Serializable {
 
     
     @Column(name="cvs_valorUnidade", precision=10)
-    public Double getCvsValorUnidade() {
+    public BigDecimal getCvsValorUnidade() {
         return this.cvsValorUnidade;
     }
     
-    public void setCvsValorUnidade(Double cvsValorUnidade) {
+    public void setCvsValorUnidade(BigDecimal cvsValorUnidade) {
         this.cvsValorUnidade = cvsValorUnidade;
     }
 
     
     @Column(name="cvs_quant")
-    public int getCvsQuant() {
+    public Integer getCvsQuant() {
         return this.cvsQuant;
     }
     
-    public void setCvsQuant(int cvsQuant) {
+    public void setCvsQuant(Integer cvsQuant) {
         this.cvsQuant = cvsQuant;
     }
